Modify rootfs.py infrastructure to support the new layout
